requirements.txt
requests
python-dotenv
gradio
pypdf
openai
openai-agents
groq
sentence-transformers
faiss-cpu
numpy